+++ app/src/main/java/com/example/moviereview/data/remote/MovieResponse.kt
@@ -1,4 +1,4 @@
-package com.example.moviereview.network
+package com.example.moviereview.data.remote
 
 data class MovieResponse(
     val display: Int,
